added job recommendation
