+++ backend-Node/routes/userRoutes.js
@@ -25,31 +25,6 @@ router.get("/get" ,(req,res) =>{
   res.send("backend working")
 })
 
-router.get("/job-recommendations", async (req, res) => {
-  const url = "https://jsearch.p.rapidapi.com/search";
-  const queryParams = {
-    query: "developer in India",
-    page: "1",
-    num_pages: "2",
-  };
 
-  const headers = {
-    "X-RapidAPI-Key": process.env.RAPIDAPI_KEY,
-    "X-RapidAPI-Host": "jsearch.p.rapidapi.com",
-  };
-
-  try {
-    const response = await axios.get(url, {
-      headers,
-      params: queryParams,
-    });
-
-    const jobs = response.data.data || [];
-    res.json({ jobs });
-  } catch (error) {
-    console.error("Error fetching jobs:", error.message);
-    res.status(500).json({ error: "Failed to fetch job recommendations" });
-  }
-});
 
 export default router
